requirements.txt
Django==2.1.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
numpy==1.16.1
pandas==0.24.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
